requirements.txt
asgiref
cffi
cryptography
dj-database-url
Django
django-rest-knox
djangorestframework
drf-yasg
inflection
packaging
pillow
psycopg2
pycparser
python-dotenv
pytz
PyYAML
setuptools
sqlparse
typing_extensions
tzdata
uritemplate
whitenoise
